Move the string-width inference into its own module
compile_string.rs answered two different questions. One is "emit the
bytecode for CONCAT", which is the job its module documentation
describes. The other is "what encoding does this string expression
have?", which arrived with #1610 and had grown to 175 lines --
string_expr_char_width and the five helpers it delegates to, plus the
tests that pin them.

That is a question about types, not about bytecode, so it moves to
string_width.rs. compile_string.rs keeps what emits, imports the one
entry point, and drops from 622 lines to 385.

A move, not a rewrite: the code is byte-identical apart from
string_expr_char_width gaining pub(crate) so its one caller can still
reach it, and the four tests move with what they test -- four #[test]
removed, the same four added.
