+++ compiler/codegen/src/lib.rs
@@ -47,6 +47,7 @@ mod emit;
 mod optimize;
 mod source_lookup;
 mod stack_balance;
+mod string_width;
 mod type_info;
 
 pub use compile::{compile, CodegenOptions};
